Rename Eyebrow to Heading, reorganize styles
- Rename `Eyebrow` component, types, and stories to `Heading`
- Move `theme.ts` to `styles/theme.ts` and update all imports
- Delete redundant CSS shim files (`styles.css`, `site-layout.css`)
- Update Storybook README title and coverage config
