+++ packages/design-system/src/components/headings/Heading.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { Typography } from '@mui/material';
-import type { EyebrowProps } from '../../types';
+import type { HeadingProps } from '../../types';
 
-function Heading({ children, tone = 'primary', as = 'span', style, ...rest }: EyebrowProps) {
+function Heading({ children, tone = 'primary', as = 'span', style, ...rest }: HeadingProps) {
   const resolvedTone = tone === 'clay' ? 'primary' : tone === 'ink' ? 'secondary' : tone;
 
   return (
@@ -28,5 +28,4 @@ function Heading({ children, tone = 'primary', as = 'span', style, ...rest }: Ey
   );
 }
 
-export { Heading as Eyebrow };
 export default Heading;
