--- packages/design-system/src/site-layout.css
@@ -1,1 +1,0 @@
-@import './styles/site-layout.css';
