+++ packages/design-system/src/stories/README.mdx
@@ -1,7 +1,7 @@
 import { Meta } from '@storybook/addon-docs/blocks';
 import { Card } from '../index';
 
-<Meta title="README" />
+<Meta title="NX°" />
 
 # Design System Storybook
 
@@ -85,7 +85,7 @@ This workspace is the visual and behavioral reference for the shared NX design-s
 
 - Brand / Logo
 - Brand / Color Palette
-- Headings / Eyebrow
+- Headings / Heading
 - Feedback / Alert
 - Feedback / Card
 - Feedback / Field
