--- packages/design-system/src/styles.css
@@ -1,1 +1,0 @@
-@import './styles/globals.css';
